ref: make 'go' canonical installer, not just alias
